fb: cfb: Introducing a command structure for framebuffer operations
Store operations for the frame buffer as commands,
Change to a method that draws at once and transfers at finalize.

Commands are managed in a linked list and processed at the finalization.
The behavior of drawing APIs does not change.
It's just that the timing at which drawing is done to memory
is internally delayed a little.

Typical usage stores command in the buffer.
The behavior of existing drawing APIs has been changed to store
commands in this buffer.
For this reason, an argument for buffer specification is added to
the initialization function cfb_display_init.

Signed-off-by: TOKITA Hiroshi <tokita.hiroshi@gmail.com>