script: gen_isr_tables: Support creating a table by function numbers.
The Renesas RA interrupt controller does not correspond to an interrupt
number and function but is configurable by the user.
Zephyr's isr_table is designed under the assumption that interrupt
numbers are assigned to specific functions, making this difficult to
support.
We will extend gen_isr_table.py to create a table by specifying the
function number instead of the interrupt number.

When creating a table with function numbers, interrupt handlers are
assigned starting from interrupt number 0 in the order they are found
during linking.
At this time, it will also create a table that allows us to reference
the assigned interrupt number from the function number.

Signed-off-by: TOKITA Hiroshi <tokita.hiroshi@gmail.com>