drivers: intc: renesas_ra: Handle interrupts based on event numbers
In the Renesas RA, interrupt numbers are not fixed because the number
and function are set in the ICU settings.

As a CUSTOM INTERRUPT CONTROLLER, we handle interrupts using event
numbers instead of interrupt numbers.

Internally, it looks up the interrupt number from the event number and
uses that number to set the interrupt for the NVIC.

Signed-off-by: TOKITA Hiroshi <tokita.hiroshi@gmail.com>