update rp2350.dtsi
SPI, I2C, ADC, and WDT work.
The `die_temp` which is subordinate to adc, also works. Added it.
The `vreg` does not exist on the rp2350. Removed it.

The configurations of gpio and die_temp are different between rp2350a and rp2350b.
I split it into each file.