soc: rp2350: Handle API change for deprecated z_arm_platform_init
https://github.com/zephyrproject-rtos/zephyr/pull/62595 introduced
changes to the kernel API. Update soc/raspberrypi/rpi_pico/rp2350/soc.c
to handle this change. If the PR isn't going to be reverted, then this
commit should be squashed into 50526fc6182 , but defer that until
https://github.com/zephyrproject-rtos/zephyr/issues/78386 has been
commented on, and leave this as a distinct commit that might be
reverted.

Signed-off-by: Andrew Featherstone <andrew.featherstone@gmail.com>