tests: gpio_api_1pin: rpi_pico2: Skip input test to avoid errata
This test's assumption that a GPIO connected to a LED can be used as
both an output and input. This isn't true, e.g. in the case of the
Raspberry Pico 2, where reconfiguring the output as an input results in
the test hitting the silicon errata RP2350-E9.

It doesn't seem possible to map a ZTEST_SUITE in C with the contents of
a testcase.yaml, which might make for a more elegant way to describe the
limitation. This is sufficient for now.

Signed-off-by: Andrew Featherstone <andrew.featherstone@gmail.com>